Correct uses of kFullServerCard in credit_card_access_manager_unittest.cc
Previously the CreateServerCard helper took a `masked` parameter
which could be used to create a kFullServer card and add it
to the TestPaymentsDataManager. However, this is not correct
behavior nowadays - PaymentsDataManager (and its test equivalent)
should not hold FullServerCards, only LocalCards or
MaskedServerCards.

This CL corrects the situation, by changing most callsites in
the test to use a masked server card instead, or in a few cases
by explicitly creating a masked server card and then un-masking
it into a full server card (where the test requires this).

Bug: 40939195

Change-Id: Icd79c95ce9f10088047015db361c030460f340c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5829086
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Olivia Saul <jsaul@google.com>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349563}