Clean up kCustomWebSignInInterceptForSupervisedUsers flag.
This flag was used to control the infrastructure changes for the custom
web sign-in intercept for supervised users. The flag has been enabled by
default since M128.

Note that the flag for the UI change
kCustomWebSignInInterceptForSupervisedUsersUi is not enabled by default,
and is not cleaned up in this CL.

Change-Id: I732231311d31fb1ca779cef74cb8329e7f63cead
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5829655
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: James Lee <ljjlee@google.com>
Reviewed-by: Monica Basta <msalama@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349779}