[ios] Mark AboutThisSiteServiceFactory as created with Profile
Manually creating the service in DoFinalInitForServices() was a
work-around for unsafety in ServiceIsCreatedWithBrowserState().

Since the unsafety in ServiceIsCreatedWithBrowserState() has been
fixed with https://crrev.com/c/5558505, it is now the recommended
way to have the service created with the Profile.

Bug: none
Change-Id: I2c18d8f73ac0b01a02c8ded051a3f81abc17af50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823065
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1349850}