[iOS][LastVisited] Deflaking testLastVisitedSectionDisplaysToday
`testLastVisitedSectionDisplaysToday` added a previous visit which took
place one hour ago from the current time. The flakiness arose when the
test was ran within one hour from midnight, as the timestamp of the
previous visit was 'Yesterday', rather than 'Today'. To fix this, this
CL makes the previous visit to take place one day ago from the current
time, making sure that 'Yesterday' is always displayed. Furthermore, the
name of the test was adjusted to match the new logic.

Bug: 363916339
Change-Id: Ieeae876b3d84a300aaa7c87703124337a04c5c80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5830767
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Filipa Senra <fsenra@google.com>
Commit-Queue: Cristian Ciacu <cristianciacu@google.com>
Cr-Commit-Position: refs/heads/main@{#1349852}