kiosk: Introduce load_profile_unittest
This also caught a bug in load_profile where the new test proved a
base::Unretained(this) that was assumed to be safe was actually unsafe
(shocking). The bug is fixed by using weak pointers in the callbacks
passed from ProfileLoader to each of the job steps.

Fixed: 290172046
Test: testing/xvfb.py tools/autotest.py -C out/Release chrome/browser/ash/app_mode/load_profile_unittest.cc
Change-Id: I6df0ab45900945ab31cd89b4832b4e19cc0a6ed0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5773922
Reviewed-by: Sergii Bykov <sbykov@google.com>
Commit-Queue: Edman Anjos <edman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349859}