[ios] Mark ListFamilyMembersServiceFactory as created with Profile
Manually creating the service in DoFinalInitForServices() was a
work-around for unsafety in ServiceIsCreatedWithBrowserState().

Since the unsafety in ServiceIsCreatedWithBrowserState() has been
fixed with https://crrev.com/c/5558505, it is now the recommended
way to have the service created with the Profile.

Bug: none
Change-Id: I04de47d084c99d57337362c470d3694e99ee84d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5831429
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1349889}