Simplify AutofillCounter
AutofillCounter is used to count the number of autocomplete entries,
addresses and credit cards stored by Autofill for the purposes of
browsing history removal. Currently, the code directly depends on
AutofillWebDataService to query Autofill's database.
For autocomplete entries, this is necessary, since Autofill doesn't
maintain an in-memory cache of autocomplete entries - they are queried
on demand at suggestion generation time. But for addresses and cards,
querying the database directly is unnecessary. The PersonalDataManager
maintains an in-memory cache of all this data. Using it is simpler,
because it removes the asynchronicity.

This CL injects the PersonalDataManager into the AutofillCounter through
the constructor and removes the asynchronous address and credit card
counting logic in favor of PersonalDataManager function calls.
Not only does this simplify the code, but it also makes it easier for
Autofill folks to maintain, since AutofillWebDataService should really
be an implementation detail that external code shouldn't depend on.

Bug: 354706653, 40100455
Change-Id: I80ece915558be5eeed7cb5f3d1dcb207afed6d72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823927
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349890}