[iOS] Re-enable testPlusAddressFallback egTest
FakePlusAddressService is used instead of an instance of the
PlusAddressService for the egTests.

In a nutshell, when a plus address is created, PlusAddressService informs this to a backend server that saves this plus address for this particular user and domain.
On get calls, this plus address is fetched from the backend by the PlusAddressService.
In the egTest, we want to make sure that there is an existing plus address saved to test out the manual fill behavior.
Using FakePlusAddressService seemed to be a simpler and better solution.

Bug: 327838014
Change-Id: I88e5a5a4431e4bdae64e3d6ed5753ac9d8aff262
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823620
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1349908}