[iOS] Record user actions in the AccountsTVC
This should record the same actions in LegacyAccountsTVC. And no changes
are expected to these metrics.

Also implements SettingsControllerProtocol in the new controller to
record dismissal.

Note: AccountDetails part of the naming exists for historical reasons,
maybe it is worth removing it in the future.

Fixed: 349071402
Change-Id: Ie86f3e57dc8c36c476665cb2456d418b85e98dea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5828962
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1349917}