[PA] Split SlotSpanMetadata into read-only one and writable one.
This is a part of ShadowMetadata CL (4/5). The entire CL is
crrev.com/c/5570783.

SlotSpanMetadata<kReadOnly> is used for SlotSpanMetadata inside
the giga cage, and SlotSpanMetadata<kWritable> is used for
SlotSpanMetadata inside the shadow pool.

PS18 = baseline.

The design doc of shadow metadata is
https://docs.google.com/document/d/1LvMI_c1VuT4sWUqgKOdJlrZv0UWP5e3PLJ_oOlJ4l2k

Bug: 40238514
Change-Id: I979a55bb35c0c186d882334931219dea80bb6538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5684977
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Ali Hijazi <ahijazi@chromium.org>
Commit-Queue: Takashi Sakamoto <tasak@google.com>
Cr-Commit-Position: refs/heads/main@{#1349950}