Reduce fragility in SidePanelCoordinator (part 1 / N)
This CL is a refactor with no intended behavior change.

This CL modifies the Show() method to take a UniqueKey rather than a
SidePanelEntry. This is in preparation for not storing a SidePanelEntry
member. As this CL audited all callsites, we can now enforce that the
SidePanelEntry looked up from UniqueKey is not nullptr.

Change-Id: I84240def2f1a11c84631e52a7c2a4b6531f6335c
Bug: 363743081
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5828329
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349957}