Replace IsNextAllocationSynchronous by AllocatePlusAddressSynchronously
After some more thought, this CL replaces a getter that would indicate
whether the callback passed to AllocatePlusAddress is run
synchronously. Instead, it exposes a method that returns a
std::optional<PlusProfile> and thus encodes in the function signature
that the operation is synchronous.

Neither approach is perfect, since in both cases the new method needs
to be kept in sync with the error checking logic in the asynchronous
allocation method.

Bug: 362445807
Change-Id: Ie97e8934f1bb20dd049dc907037cde9b32cbf0da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5829215
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1350014}