Add button actions to prediction improvements suggestion
This CL wires up the acceptance of a button action in
kFillPredictionImprovements suggestion to the delegate.
For now, the functionality is missing.

Please not that there is a chance that a new suggestion type will be
responsible for handling this actions. This will depend on whether we
decide that the feedback suggestion will have its own suggestion type
or not.

Bug: b/362468426
Change-Id: I5fb5c173828cf6a38e9605fd0397fddb9ff0dbda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5822458
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Wolfgang Billenstein <bwolfgang@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1350043}