[UPMAccessLossWarning] Add help center button
This CL introduces the custom view for access loss dialog in settings.
Because the standard dialog doesn't allow to place an image to the right
of the title, new custom view was needed. Now it sets title, description
and the help button visibility on the custom view model and sets this
custom view on the modal dialog model.

Bug: 361286381
Change-Id: Ief2d964fbdd6f729109e32027f976ac8c551a05b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826667
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350049}