Make all AppDefinedDomain read async and cached
The AppDefinedDomains are read from a system service, which may have a
significant delay. On the other hand, since these are values from the
apps manifest, they are not expected to change during the runtime of the
app.

This CL ensures that the information is always read on a background
thread, as it was already done by the
masked_domain_list_component_loader.
It also caches results in memory to speed up subsequent calls.

It is possible that GetAppDefinedDomains may be called multiple times
for the same criteria before the first callback returns with a result
and caches it. In this case the following calls will overwrite the
cached value. This is not a problem per se, but slightly inefficient,
and will be fixed in a follow-up CL.

Bug: 353556455
Change-Id: Ie9c2e0cdc35fbcb7e149cfb9d4158a3dd4eed52e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823049
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Peter Pakkenberg <pbirk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350067}