[UPMAccessLossWarning] Add the action for the sheet button
The SimpleNoticeSheet supports adding a Runnable that runs when the
button is clicked.
On the AccessLossWarning sheet, the buttons are wired to either
lead to GMSCore update or password export.

open to track that the actions which are taken on button click are not
tested.

Bug: 356850591, 356850882
Change-Id: I4554c9151a402c853675d6649b36ebb793c02291
Low-Coverage-Reason: hard to test, bugs b/322480299 and b/364247783 are
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823920
Reviewed-by: Friedrich Hauser <fhorschig@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1350130}