Add browser tests for auction deal support.
This change tests that when we run an auction with this feature on,
and participate with selectableBuyerAndSellerReportingIds that
within each step (generateBid, scoreAd, reportWin, reportResult),
the expected amount of reporting ids is what we expect.

The three scenarios tested are what happens when
selectableBuyerAndSellerReportingIds are present in an IG ad and:
  - One is selected
  - One is not selected
  - One is selected but invalid.

There is also a test to ensure that, if the feature is disabled,
then the selectable ids and selected can not make their way into
the auction.

Most of these functionalities have been tested within unit tests
when they were submitted, however this is just an extra step to ensure
everything works as intended, end to end.

It's important to note, that this does not test the change with
K-Anon enforcement, as the current architecture of browser tests
do not make that an easy task. Those will be tested separately in
another CL.


Bug: 356654297
Change-Id: I16562359df5cce9c354ff5303daee413485dc8f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5806111
Commit-Queue: Youssef Bourouphael <ybourouphael@google.com>
Reviewed-by: Orr Bernstein <orrb@google.com>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350171}