[views-ax] Migrating kSelection event in Views
This CL migrates the firing of the kSelection event in Views to
the ViewAccessibility setter, as part of the ViewsAX project.

This should've been included in CL:5675624 but we forgot to include
it there.

OmniboxPopupViewViews required a more involved refactoring, since
there were many different scenarios of when the selection a11y event
was being fired.

Part of the change to this was making sure that the OmniboxRowView's
buttons actually now have a a11y selected state, and as such fire the
selection events when this changes, rather than firing the event
manually from the popup view or the result view.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: Id873e774a10fd8fe28e330b5df8675e525e7ed12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5803672
Auto-Submit: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350178}