Add a source_set for "platform_browser_tests"
Tests inheriting from PlatformBrowserTest are intended to run on both
desktop and Android platforms. In order to accomplish this, they are
included separately in the targets for browser_tests and
android_browsertests.

At best, this leads to tedium by needing to specify the same file
twice, but this can also lead to the much more negative case of only
running these tests on one set of platforms (e.g., like the
tests fixed in https://crrev.com/7d976b57ce7).

Avoid the need to specify the test in two places by introducing a
new gn source_set, "platform_browser_tests", that is itself included
in both browser_tests and android_browsertests. This way, each new test
only needs to be added to this individual source_set.

This CL moves a smattering of PlatformBrowserTests to this source_set,
but not all (they are not all trivial, and there are more that are
incorrectly specified). Followups will continue moving them over.

Bug: 361002665
Change-Id: I4a07f79ae33c4b403e9f69daeeb10d9ac72c14bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825291
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350201}