Re-register the browser with the updater if the updater's existence checker path does not match.
In the case of multiple installations of the browser, this will help the
most-recently-run one be kept up-to-date.

This can also repair an issue where the existence checker path is empty,
which is fatal to the update process.

Fixed: 361707416
Change-Id: Ie9fb78b3b54730ca2df6981d9c0597a0edffba81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825292
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Auto-Submit: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350224}