Extend Fledge Virtual Test suites
These features still haven't stabilized. The deprecation trial flag is
still active and k-anonymity enforcement is not yet at 100%. We expect
to have a better idea of the timeline for removing the cookie
deprecation flag by the beginning of next year, and the timeline for the
k-anonymity roll out in 2025Q1.

Fixed: 363051807,363051808
Change-Id: I110dbadf96449a14c8c278532a9680565df496e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5833929
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1350227}