PrivateAggregation: Rename Api enum to PrivateAggregationCallerApi
This CL creates a new header file to house PrivateAggregationCallerApi,
an enum that identifies which API is calling Private Aggregation.

The enum used to be PrivateAggregationBudgetKey::Api, but now we're
using it for things that aren't related to budgeting.

Fixed: 362431899
Change-Id: Ia0e7fbe9f19119557dc4e6bb7406206bde9cfc36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823744
Reviewed-by: Cammie Smith Barnes <cammie@chromium.org>
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Alex Turner <alexmt@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350232}