Add controls to remove SwiftShader as an automatic WebGL fallback.
Add a AllowSwiftShaderFallback finch feature as a killswitch. This
is currently *enabled by default* with a deprecation message.

Once AllowSwiftShaderFallback is disabled:

Disable falling back to SwiftShader as a result of multiple GPU crashes.
Instead, fall back to the display compositor.

Disable falling back to SwiftShader as a GL implementation unless it's
the only available implementation (current behaviour) AND
--allow-unsafe-swiftshader is passed to Chrome. This allows testing of
Chrome on systems without GPUs without exposing end users to
security issues in the SwANGLE stack.

Continue to allow SwiftShader to be used if it's explicitly requested
by command line with the --use-angle=swiftshader flag. This allows testing of Chrome's rasterization and WebGL using SwiftShader.

Bug: 40277080
Change-Id: I8b6f56acba1acb7a988232fb2f205a6be0da670f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5675974
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350236}