Fix Windows build with enable_printing=false
Some policy keys and are not available when printing is disabled.
Also, the printing//BUILD.gn has an assertion that fails when
the file is included when printing is off, so ensure that this
does not trigger by adding a condition in the top level BUILD.gn.

Bug: chromium:364089367
Change-Id: I7fcc0c4660a73d9e0643245f83d19a2d253c536b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5829417
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Auto-Submit: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350238}