Prevent Reentrant behavior in PermissionRequestManager
If the object requesting a permission attempts to request a new
permission in it's result callback, then the PermissionRequestManager
may decide to preempt the previous permission request with the new
permission. However, this callback is called via an iterator on the
`requests_` object and preempting a callback clears this `requests_`
object leading to an invalid iterator and thus a crash.

As a previous workaround it had seemed like sites could post a task
to handle the new permission request; however, it seems that there may
often be a race where the new permission is requested, but the bubble
is scheduled for destruction. This causes the bubble to get destroyed
and for the new/secondary permission to be rejected.

WebXR encounters this because we need to first request the overall
permission to enter an XR session, and then if that is granted, we
want to prompt for any permissions that may be required for features
that were requested to be enabled on the session. However, if the
initial session level permission is rejected, we don't want to prompt
for permissions on any of the features.

This change removes the `PostTask` logic from WebXR that was causing
permissions to be autorejected and adds further logic to
PermissionRequestManager to prevent a request from being preempted if
it is in the final stages of being resolved (e.g. a decision was already
made).

Fixed: 357776212
Change-Id: I54c78ceec7daecba530b397f7bbb74492e57858f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825068
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1350249}