[iOS] Fix ESBInlineSettingsPromo Crash
There are reports that the ESBInlineSettingsPromo is causing the
application to crash. After investigating, the problem was identified to
be caused by the use of the BrowserState object that has been
deallocated. This CL checks the existence of the BrowerState object
before its use.

Bug: 361307172
Change-Id: I3ad701bfd1734681a4c3e73a75cda4a93bddc8f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5802379
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1350339}