[views-ax] Migrating Role to view constructor in ui/views directory
This CL migrates setting of the accessible role from
View::GetAccessibleNodeData to the view constructor. This ensures that
the role is updated in the accessibility cache as soon as the view is
constructed. The migration is applied to following files:

 1. ui/views/controls/menu/menu_item_view.cc
 2. ui/views/controls/menu/submenu_view.cc
 3. ui/views/controls/menu/menu_scroll_view_container.cc

The remaining instances will be migrated in follow-up CLs.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-

Bug: 325137417
Change-Id: I3810ff24cb61c2c53d1609fd6787e07543d4a31e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5821238
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Sejal Anand <sejalanand@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1350340}