[lensoverlay] Send PDF bytes for contextualization
When the feature flag is enabled, extracts the PDF bytes from the PDF
renderer process and sends them to the Lens server to help with
contextualization. This CL is meant to help prototype and unblock server
side work, not be used on any real users yet. Therefore the logic is
gated by a feature flag defaulted to off. Bytes only leave the device if
the user invokes the Lens Overlay after manually turning on the feature
flag.

Before enabling the flag for real users, updates will be made to our
privacy data notice so no bytes leave the device unless the user
consents.

More info in the design doc linked in the bug.

Bug: 361409401
AX-Relnotes: n/a.
Change-Id: I9c261fe4e02a607a845bcef6bb2375830880ff37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5811891
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#1350343}